Core/Spell : implement gladiator stance

+++ src/server/game/Spells/SpellInfo.h
@@ -333,8 +333,8 @@ public:
     uint32 AttributesEx11;
     uint32 AttributesEx12;                                  // new 5.4.0
     uint32 AttributesCu;
-    uint32 Stances;
-    uint32 StancesNot;
+    uint64 Stances;
+    uint64 StancesNot;
     uint32 Targets;
     uint32 TargetCreatureType;
     uint32 RequiresSpellFocus;
